grades: Move Class to internal directory
