+++ src/grades/grades.pro
@@ -5,15 +5,15 @@ CONFIG += staticlib
 QML_IMPORT_PATH += qml
 
 SOURCES += \
-    Class.cpp \
     GradesModule.cpp \
+    internal/Class.cpp \
     internal/ClassManager.cpp \
     internal/GradeCalculator.cpp
 
 HEADERS += \
-    Class.h \
     GradesModule.h \
     IClassManager.h \
+    internal/Class.h \
     internal/ClassManager.h \
     internal/GradeCalculator.h
